+++ src/store/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+
+
+const reducers = combineReducers({})
+
+export default reducers;
